apps/processor/microservices/units/ms-atm-pdf/pyproject.toml
[project]
name = "ms-atm-pdf"
version = "0.1.0"
description = "PDF/OCR Atomizer microservice – extracts text and tables from PDF files with OCR fallback."
requires-python = ">=3.12"
dependencies = [
    # PDF processing
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    # OCR
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    # gRPC
    "grpcio>=1.62",
    "grpcio-tools>=1.62",
    "grpcio-health-checking>=1.62",
    # FastAPI (for health endpoint)
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    # HTTP client
    "httpx>=0.27",
    # Pydantic
    "pydantic>=2.0",
    # OpenTelemetry
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "opentelemetry-instrumentation-grpc>=0.44b0",
    "opentelemetry-exporter-otlp>=1.23",
    # Dapr
    "dapr>=1.13",
    "dapr-ext-grpc>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
